Ooops.  ssize_t != int.  <sigh>
Thanks to: Oliver Lehmann and Peter Wemm


git-svn-id: svn+ssh://svn.freebsd.org/base/head@137296 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ lib/libarchive/archive_write_set_compression_bzip2.c
@@ -57,7 +57,7 @@ struct private_data {
 
 static int	archive_compressor_bzip2_finish(struct archive *);
 static int	archive_compressor_bzip2_init(struct archive *);
-static ssize_t	archive_compressor_bzip2_write(struct archive *, const void *,
+static int	archive_compressor_bzip2_write(struct archive *, const void *,
 		    size_t);
 static int	drive_compressor(struct archive *, struct private_data *,
 		    int finishing);
